+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Renamed wasped options -set and -unset to --verify and --unverify
+
 ## [0.5.0] - 2021-02-11
 
 - Summary shows more clearly number of requirements left to verify